update approve page

+++ workspace/ims-webapp/src/main/java/th/co/baiwa/excise/ia/persistence/repository/RentHouseRepository.java
@@ -1,8 +1,10 @@
 package th.co.baiwa.excise.ia.persistence.repository;
 
+import java.math.BigDecimal;
+
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.co.baiwa.excise.ia.persistence.entity.RentHouse;
 
-public interface RentHouseRepository extends CommonJpaCrudRepository<RentHouse, Long> {
+public interface RentHouseRepository extends CommonJpaCrudRepository<RentHouse, BigDecimal> {
 
 }
